#74: Move SymmetryChecker into the cpp file, and replace with a global function check_symmetry()
